amplify/functions/src/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
boto3
pydantic>=2.0.0
pydantic-core>=2.0.0
pydantic-settings